+++ application/controllers/link.php
@@ -8,16 +8,45 @@ class Link extends MY_Controller {
 	function __construct() {
 	    parent::__construct();
 
-		if ($this->userdata['group']['name'] != 'admin')
-		{
-			$this->session->set_flashdata('error', 'You do not have appropriate permissions for this action.');
-			redirect('feed/page');
-		}
+		$this->load->model('link_model');
 	}
 
-	public function view()
+	public function search()
 	{
-		$this->load->model('link_model');
+		if ($this->uri->segment(4))
+		{
+			$data['table'] = $this->link_model->read(array('where' => array($this->uri->segment(3) => $this->uri->segment(4))));
+		}
+		else
+		{
+			$data['table'] = $this->link_model->read();
+		}
+
+		//print_r($data);
+
+		$data['title'] = 'Links';
+		$data['backtrack'] = array('resource' => 'Resources', 'link' => 'Links', 'link/' => 'View');
+
+
+	    // print the page
+		$this->output->set_header("Cache-Control: max-age=300, public, must-revalidate");
+
+		$this->load->view('head', array('page_title' => $data['title'], 'stylesheets' => array('demo_table.css', 'layout_outer.css', 'layout_inner.css', 'theme.css'), 'scripts' => array('jquery.dataTables.min.js', 'link.js')));
+		$this->load->view('header');
+		$this->load->view('main_open');
+		$this->load->view('left_column');
+		$this->load->view('link_list_view', $data);
+		$this->load->view('main_close');
+		$this->load->view('footer');
+	}
+
+	public function admin()
+	{
+		if (! $this->userdata['is_admin'])
+		{
+			$this->session->set_userdata('error', 'You do not have appropriate permissions to administer links.');
+			redirect('link');
+		}
 
 		if ($this->uri->segment(4))
 		{
@@ -48,14 +77,20 @@ class Link extends MY_Controller {
 
 	public function create()
 	{
+		if (! $this->userdata['is_admin'])
+		{
+			$this->session->set_userdata('error', 'You do not have appropriate permissions to administer links.');
+			redirect('link');
+		}
+
 	    $this->load->helper(array('form', 'url'));
-	    $this->load->library(array('form_validation', 'link_class'));
-		$this->load->model('link_model');
+	    $this->load->library(array('form_validation', 'link_class', 'tag_class'));
 
 		$this->form_validation->set_error_delimiters('<div class="error">', '</div>');
 
 	    $this->form_validation->set_rules('url', 'URL', 'required');
 	    $this->form_validation->set_rules('title', 'Title', 'required');
+	    $this->form_validation->set_rules('tags', 'Tags', 'callback_tag_check');
 
 		if ($this->form_validation->run() == false)
 		{
@@ -79,7 +114,7 @@ class Link extends MY_Controller {
 		    if ($id = $this->link_class->create($this->input->post()))
 		    {
 				$this->session->set_flashdata('message', 'Link successfully created.');
-		        redirect('link/view/id/'.$id);
+		        redirect('link/search/id/'.$id);
 		    }
 		    else
 		    {
@@ -90,6 +125,12 @@ class Link extends MY_Controller {
 
 	public function edit()
 	{
+		if (! $this->userdata['is_admin'])
+		{
+			$this->session->set_userdata('error', 'You do not have appropriate permissions to administer links.');
+			redirect('link');
+		}
+
 	    if ($this->uri->segment(3, null) == null && $this->input->post('id') == null)
 	    {
 			$this->session->set_flashdata('error', 'You must specify a user to edit, or simply create a new one here. [007]');
@@ -97,19 +138,19 @@ class Link extends MY_Controller {
 	    }
 
 	    $this->load->helper(array('form', 'url'));
-	    $this->load->library(array('form_validation', 'link_class'));
-		$this->load->model('link_model');
+	    $this->load->library(array('form_validation', 'link_class', 'tag_class'));
 
 		$this->form_validation->set_error_delimiters('<div class="error">', '</div>');
 
 	    $this->form_validation->set_rules('url', 'URL', 'required');
 	    $this->form_validation->set_rules('title', 'Title', 'required');
+	    $this->form_validation->set_rules('tags', 'Tags', 'callback_tag_check');
 
 		if ($this->form_validation->run() == false)
 		{
 			$data = $this->link_class->full_form($this->uri->segment(3));
 
-	        $data['target'] = 'edit';
+	        $data['target'] = 'edit/'.$this->uri->segment(3);
 			$data['form_title'] = 'Edit Link Info';
 			$data['controls'] = anchor('link/view/id/'.$this->uri->segment(3), img(base_url().'img/cancel_icon.png'), array('class' => 'cancel'));
 			$data['backtrack'] = array('resource' => 'Resources', 'link/view' => 'Links', 'link/edit/'.$this->uri->segment(3) => 'Edit');
@@ -127,7 +168,7 @@ class Link extends MY_Controller {
 		    if ($id = $this->link_class->edit($this->input->post()))
 		    {
 			    $this->session->set_flashdata('success', 'Link successfully edited.');
-			    redirect('link/view/id/'.$id);
+			    redirect('link/search/id/'.$id);
 			}
 		    else
 		    {
@@ -138,6 +179,12 @@ class Link extends MY_Controller {
 
 	public function delete()
 	{
+		if (! $this->userdata['is_admin'])
+		{
+			$this->session->set_userdata('error', 'You do not have appropriate permissions to administer links.');
+			redirect('link');
+		}
+
 		if (! $this->link_model->delete($this->uri->segment(3, null)))
 		{
 			$error = "Couldn't delete link.";
@@ -145,6 +192,20 @@ class Link extends MY_Controller {
 		if (isset($error)) {
 			$this->session->set_flashdata('error', $error);
 		}
-		redirect('link/view');
+		redirect('link');
+	}
+
+	// validation callback function for checking tags
+	function tag_check($tags)
+	{
+		if ($fail = $this->tag_class->check_tag_input($tags))
+		{
+			$this->form_validation->set_message($fail[0], $fail[1]);
+			return false;
+		}
+		else
+		{
+			return true;
+		}
 	}
 }
